AI-Powered-Project-Management-Application/backend/requirements.txt
Flask==2.3.3
flask-cors==4.0.0
psycopg2-binary==2.9.9
python-dotenv==1.0.0
PyJWT==2.8.0
bcrypt==4.0.1
reportlab==4.0.7
requests==2.31.0
python-jose==3.3.0
Werkzeug==2.3.7
gunicorn==21.2.0
python-dateutil==2.8.2 